Implemented some Guild management, including allowing the guildmaster to approve and deny applications.
Fixed a holdover from my OleDB days where I didn't use parameters to insert text into the database, which required the use of an extra apostrophe to escape it.

Once again fixed the correct display of Enter and Apply buttons on the Guild List Page.

Corrected the text that displays when you sleep at the Guild.

Fixed an issue where it didn't update the list of Members correctly if you expelled one as a guildmaster.