The Enemy can now flee during battles if they are under 40% endurance.
I had intended for it to be under 20%, but I decided to give them a chance at 40%. At 10%, they have a much higher chance of fleeing, or if their level is much lower in comparison to the Player.

I made a list of reserved names that cannot be chosen to cut down on repetitive !string.Equals calls.

I fixed an issue where I forgot to add my Thirst/HungerToStringWithText calls to the NotifyPropertyChanged in the Hunger/Thirst Properties.

I fixed a possible issue where my Bonus() method in the BattlePage was not necessarily calculating the right result due to operator precendence.